requirements.txt
Django==2.2.1
channels==2.2.0
google-cloud-translate==1.4.0
channels_redis==2.4.0
names==0.3.0